feat(nat): add common module and unify nat actions
Let masquerading and port forwarding use the same type for actions.
