+++ nat/src/portfw/nf.rs
@@ -16,7 +16,7 @@ use std::num::NonZero;
 use std::sync::Arc;
 use std::time::Instant;
 
-use crate::portfw::flow_state::NatAction;
+use crate::common::NatAction;
 use crate::portfw::flow_state::build_portfw_flow_keys;
 use crate::portfw::flow_state::get_packet_port_fw_state;
 use crate::portfw::flow_state::refresh_port_fw_entry;
